Changes as per discussions in SWiG Main
Changed ID from an integer value with a maximum of 255 (8 bits) to an integer with a maximum of 65535 (16 bits) to permit vendor and technology-specific ID groups, as well as preventing running out of room.

Added a "null" type for a parameter. This allows creation of a parameter that can be used solely for linking atomic parameters without specifying something of its own